# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Kiroulou API Documentation",
        "description": "Documentation de l'API utilisée par l'application mobile KiRoulOu, faite avec Laravel",
        "contact": {
            "email": "bezedache29@gmail.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "0.0.1"
    },
    "servers": [
        {
            "url": "https://localhost/api",
            "description": "Server API KiRoulOu"
        }
    ],
    "paths": {
...
        "/users/posts": {
            "post": {
                "tags": [
                    "Users"
                ],
                "summary": "Create user post",
                "description": "Ajout d'un article par un user",
                "operationId": "c9638d9064495990bcf7c353a37bed95",
                "requestBody": {
                    "$ref": "#/components/requestBodies/AddPostUser"
                },
                "responses": {
                    "201": {
                        "description": "Article d'un user créé",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "post created"
                                        },
                                        "post_user_id": {
                                            "type": "string",
                                            "example": "e6c64643-6ac9-4adb-87d2-bfe121dc4580"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "$ref": "#/components/responses/UnprocessableEntity"
                    }
                },
                "security": [
                    {
                        "bearer_token": []
                    }
                ]
            }
        },
...
        "/users/posts/{post_id}/likeOrUnlike": {
            "post": {
                "tags": [
                    "Users"
                ],
                "summary": "Like or unlike user post",
                "description": "Aimer ou ne plus aimer un article d'un user",
                "operationId": "ed2a38ddc7e5324fb592603f9dd0c774",
                "parameters": [
                    {
                        "$ref": "#/components/parameters/post_id"
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Article d'un user créé",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "liked"
                                        },
                                        "post": {
                                            "$ref": "#/components/schemas/PostUserCounts"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "$ref": "#/components/responses/UnprocessableEntity"
                    }
                },
                "security": [
                    {
                        "bearer_token": []
                    }
                ]
            }
        },
        "/users/posts/{post_id}/comments": {
            "post": {
                "tags": [
                    "Users"
                ],
                "summary": "Create user post comment",
                "description": "Ajout d'un commentaire à un article d'un user",
                "operationId": "c676f7433626c198ebb1f1fbdc48422e",
                "parameters": [
                    {
                        "$ref": "#/components/parameters/post_id"
                    }
                ],
                "requestBody": {
                    "$ref": "#/components/requestBodies/AddPostUserComment"
                },
                "responses": {
                    "201": {
                        "description": "Commentaire d'article de user créé",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "comment created"
                                        },
                                        "comments": {
                                            "$ref": "#/components/schemas/PostUserComment"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "$ref": "#/components/responses/UnprocessableEntity"
                    }
                },
                "security": [
                    {
                        "bearer_token": []
                    }
                ]
            }
        },
        "/users/posts/{post_id}/comments?page={page}": {
            "get": {
                "tags": [
                    "Users"
                ],
                "summary": "All user's posts",
                "operationId": "94d8b728b19cc11a03ab58ee72347962",
                "parameters": [
                    {
                        "$ref": "#/components/parameters/post_id"
                    },
                    {
                        "$ref": "#/components/parameters/page"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/PostUserComment"
                                    }
                                }
                            }
                        }
                    },
                    "404": {
                        "$ref": "#/components/responses/NotFound"
                    }
                },
                "security": [
                    {
                        "bearer_token": []
                    }
                ]
            }
        }
    },
    "components": {
        "schemas": {
            "Address": {
                "description": "Adresse simplifié",
                "properties": {
                    "street_address": {
                        "description": "Rue / lieu-dit",
                        "type": "string",
                        "example": "Ma super adresse"
                    },
                    "lat": {
                        "description": "Lattitude",
                        "type": "string",
                        "example": "48.5740185"
                    },
                    "lng": {
                        "description": "Longitude",
                        "type": "string",
                        "example": "-4.3335965"
                    },
                    "region": {
                        "description": "Region",
                        "type": "string",
                        "example": "Ma super région"
                    },
                    "department": {
                        "description": "Département",
                        "type": "string",
                        "example": "Mon super département"
                    },
                    "department_code": {
                        "description": "Code du département",
                        "type": "string",
                        "example": "29"
                    },
                    "city": {
                        "$ref": "#/components/schemas/City"
                    },
                    "zipcode": {
                        "$ref": "#/components/schemas/Zipcode"
                    }
                },
                "type": "object"
            },
...
            "City": {
                "description": "Ville",
                "properties": {
                    "name": {
                        "type": "string",
                        "example": "Lesneven"
                    }
                },
                "type": "object"
            },
...
            "PostUserSimple": {
                "description": "Article d'un user simple avec les compteurs",
                "properties": {
                    "id": {
                        "type": "string",
                        "example": "b7d15d52-0458-4a39-a59e-a7a370e7f31c"
                    },
                    "title": {
                        "type": "string",
                        "example": "Mon super titre"
                    },
                    "description": {
                        "type": "string",
                        "example": "Ma super description d'article"
                    },
                    "user_id": {
                        "type": "number",
                        "example": 10005
                    },
                    "created_at": {
                        "type": "string",
                        "example": "2022-08-15 14:25:01"
                    },
                    "updated_at": {
                        "type": "string",
                        "example": null
                    },
                    "deleted_at": {
                        "type": "string",
                        "example": null
                    }
                },
                "type": "object"
            },
...
            "PostUserCounts": {
                "description": "Article d'un user avec les compteurs de commentaires et likes",
                "type": "object",
                "allOf": [
                    {
                        "$ref": "#/components/schemas/PostUserSimple"
                    },
                    {
                        "properties": {
                            "post_user_likes_count": {
                                "type": "number",
                                "example": 12
                            },
                            "post_user_comments_count": {
                                "type": "number",
                                "example": 3
                            }
                        }
                    }
                ]
            },
...
            "PostUserComment": {
                "description": "Commentaire d'un article d'un user par un user",
                "properties": {
                    "id": {
                        "type": "number",
                        "example": 1
                    },
                    "user_id": {
                        "type": "number",
                        "example": 10005
                    },
                    "post_user_id": {
                        "type": "string",
                        "example": "0d2991f8-132b-43fc-87f9-9bbcbf0d7877"
                    },
                    "message": {
                        "type": "string",
                        "example": "Mon super commentaire d'un article d'un user"
                    },
                    "created_at": {
                        "type": "string",
                        "example": "2022-08-15 14:25:01"
                    },
                    "updated_at": {
                        "type": "string",
                        "example": null
                    },
                    "deleted_at": {
                        "type": "string",
                        "example": null
                    },
                    "user": {
                        "$ref": "#/components/schemas/UserDetails"
                    }
                },
                "type": "object"
            },
...
            "User": {
                "description": "User Simple",
                "properties": {
                    "id": {
                        "description": "Id du user",
                        "type": "number",
                        "example": "1"
                    },
                    "email": {
                        "description": "Email du user",
                        "type": "string",
                        "example": "simon-strueux@gmail.com"
                    },
                    "created_at": {
                        "description": "Date de création du user",
                        "type": "string",
                        "format": "date-time",
                        "nullable": true
                    },
                    "updated_at": {
                        "description": "Date d'édition du user",
                        "type": "string",
                        "format": "date-time",
                        "nullable": true
                    },
                    "deleted_at": {
                        "description": "Date d'e suppression du user",
                        "type": "string",
                        "format": "date-time",
                        "nullable": true
                    }
                },
                "type": "object"
            },
...
            "UserDetails": {
                "description": "Détails du user connecté",
                "type": "object",
                "allOf": [
                    {
                        "$ref": "#/components/schemas/User"
                    },
                    {
                        "properties": {
                            "firstname": {
                                "description": "Prénom du user",
                                "type": "string",
                                "example": "Simon",
                                "nullable": true
                            },
                            "lastname": {
                                "description": "Nom du user",
                                "type": "string",
                                "example": "Strueux",
                                "nullable": true
                            },
                            "is_push_notifications": {
                                "description": "Est-ce que le user souhaite recevoir des notifications push",
                                "type": "number",
                                "example": 1
                            },
                            "is_email_notifications": {
                                "description": "Est-ce que le user souhaite recevoir des notifications email",
                                "type": "number",
                                "example": 1
                            },
                            "club_id": {
                                "description": "membre du club ayant l'id du club_id",
                                "type": "number",
                                "example": 1
                            },
                            "is_club_admin": {
                                "description": "Est-ce que le user est admin de son club",
                                "type": "number",
                                "example": 1
                            },
                            "club_name": {
                                "description": "Nom du club, si le user est dans un club",
                                "type": "string",
                                "example": "Côte Des Légendes VTT"
                            },
                            "premium_name": {
                                "description": "Nom de l'abonnement du user, s'il est abonné",
                                "type": "string",
                                "example": "Premium 1"
                            },
                            "address": {
                                "$ref": "#/components/schemas/Address"
                            }
                        }
                    }
                ]
            },
...
            "Zipcode": {
                "description": "Code postal",
                "properties": {
                    "code": {
                        "type": "string",
                        "example": "29260"
                    }
                },
                "type": "object"
            }
        },
        "responses": {
...
            "UnprocessableEntity": {
                "description": "Ressource non traitée",
                "content": {
                    "application/json": {
                        "schema": {
                            "properties": {
                                "email": {
                                    "type": "array",
                                    "items": {},
                                    "example": [
                                        "L'adresse email est obligatoire"
                                    ]
                                }
                            },
                            "type": "object"
                        }
                    }
                }
            },
...
            "NotFound": {
                "description": "Ressource introuvable",
                "content": {
                    "application/json": {
                        "schema": {
                            "properties": {
                                "message": {
                                    "type": "string",
                                    "example": "Resource not found"
                                }
                            },
                            "type": "object"
                        }
                    }
                }
            }
        },
        "parameters": {
...
            "post_id": {
                "name": "post_id",
                "in": "path",
                "description": "ID du post (uuid)",
                "required": true,
                "schema": {
                    "type": "string"
                }
            },
...
            "page": {
                "name": "page",
                "in": "path",
                "description": "Numéro de la page",
                "required": true,
                "schema": {
                    "type": "integer"
                }
            },
...
        },
        "requestBodies": {
...
            "AddPostUser": {
                "required": true,
                "content": {
                    "application/json": {
                        "schema": {
                            "required": [
                                "title, description, user_id"
                            ],
                            "properties": {
                                "title": {
                                    "type": "string",
                                    "example": "Mon super titre"
                                },
                                "description": {
                                    "type": "string",
                                    "example": "Ma super description"
                                }
                            },
                            "type": "object"
                        }
                    }
                }
            },
            "AddPostUserComment": {
                "required": true,
                "content": {
                    "application/json": {
                        "schema": {
                            "required": [
                                "message",
                                "post_user_id",
                                "user_id"
                            ],
                            "properties": {
                                "message": {
                                    "type": "string",
                                    "example": "Mon super commentaire"
                                }
                            },
                            "type": "object"
                        }
                    }
                }
            }
        },
        "securitySchemes": {
            "bearer_token": {
                "type": "http",
                "description": "S'authentifier avec email & password pour récupérer un token d'authentification",
                "name": "Authorization",
                "in": "header",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}